\name{aloha}
\docType{data}
\alias{aloha}
\title{Results of the simulation of the Aloha network}
\description{
  This data set contains the results of the simulation of the Aloha network,
  that can be found in the samples directory of the OMNeT++ installation.
}
\usage{aloha}
\format{An OMNeT++ dataset containing 348 scalars, 3 vectors, and 88 statistics.}
\keyword{datasets}